#417 added ParameterFilter to enable filtering out some operation parameters in output doc. Also, did some REFACTORING as ModelController was extracted
