+++ src/main/java/com/github/kongchen/swagger/docgen/spring/ParameterFilter.java
@@ -1,0 +1,6 @@
+package com.github.kongchen.swagger.docgen.spring;
+
+public interface ParameterFilter {
+
+    boolean isIgnoredParameter(ParameterMetadata parameterMetadata);
+}
