+++ src/main/java/com/github/kongchen/swagger/docgen/spring/SpringMvcApiReader.java
@@ -4,26 +4,19 @@ import com.github.kongchen.swagger.docgen.util.LogAdapter;
 import com.github.kongchen.swagger.docgen.mavenplugin.ApiSource;
 import com.github.kongchen.swagger.docgen.util.Utils;
 import com.wordnik.swagger.annotations.Api;
-import com.wordnik.swagger.annotations.ApiModel;
-import com.wordnik.swagger.annotations.ApiModelProperty;
 import com.wordnik.swagger.annotations.ApiOperation;
 import com.wordnik.swagger.annotations.ApiResponse;
 import com.wordnik.swagger.annotations.ApiResponses;
 import com.wordnik.swagger.config.SwaggerConfig;
 import com.wordnik.swagger.converter.OverrideConverter;
-import com.wordnik.swagger.core.SwaggerSpec;
-import com.wordnik.swagger.model.AllowableListValues;
 import com.wordnik.swagger.model.ApiDescription;
 import com.wordnik.swagger.model.ApiListing;
 import com.wordnik.swagger.model.Authorization;
 import com.wordnik.swagger.model.AuthorizationScope;
 import com.wordnik.swagger.model.Model;
-import com.wordnik.swagger.model.ModelProperty;
-import com.wordnik.swagger.model.ModelRef;
 import com.wordnik.swagger.model.Operation;
 import com.wordnik.swagger.model.Parameter;
 import com.wordnik.swagger.model.ResponseMessage;
-import org.codehaus.jackson.annotate.JsonIgnore;
 import org.codehaus.plexus.util.StringUtils;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -31,16 +24,9 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import scala.Option;
 import scala.collection.JavaConversions;
-import scala.collection.mutable.LinkedHashMap;
 
-import javax.xml.bind.JAXBElement;
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlTransient;
 import java.lang.annotation.Annotation;
-import java.lang.reflect.Field;
 import java.lang.reflect.Method;
-import java.lang.reflect.ParameterizedType;
-import java.lang.reflect.Type;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
@@ -49,6 +35,9 @@ import java.util.Comparator;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+
+import static com.github.kongchen.swagger.docgen.spring.IgnoredClassParameterFilter.SPRING_IGNORED_PARAMETER_TYPES;
+import static com.github.kongchen.swagger.docgen.spring.SpringReaderUtils.getGenericSubtype;
 
 /**
  * @author tedleman
@@ -65,28 +54,31 @@ import java.util.Map;
  */
 public class SpringMvcApiReader {
     private final ParameterGenerator parameterGenerator;
-    private ApiSource apiSource;
+    private final ApiSource apiSource;
+    private final ModelCollector modelCollector;
     private ApiListing apiListing;
     private String resourcePath;
     private List<String> produces;
     private List<String> consumes;
-    private HashMap<String, Model> models;
+    private final ParameterFilter parameterFilter;
+    private final LogAdapter logger;
 
     private static final Option<String> DEFAULT_OPTION = Option.empty(); //<--comply with scala option to prevent nulls
-    private static final String[] RESERVED_PACKAGES = {"java", "org.springframework"};
-    private final LogAdapter logger;
-    private OverrideConverter overriderConverter;
 
-    public SpringMvcApiReader(ApiSource aSource, LogAdapter logger, OverrideConverter overrideConverter) {
-        apiSource = aSource;
-        this.logger = logger;
+    public SpringMvcApiReader(ApiSource aSource, OverrideConverter overrideConverter, LogAdapter logger) {
+        this(aSource, new IgnoredClassParameterFilter(SPRING_IGNORED_PARAMETER_TYPES), overrideConverter, logger);
+    }
+
+    public SpringMvcApiReader(ApiSource apiSource, ParameterFilter parameterFilter, OverrideConverter overrideConverter, LogAdapter logger) {
+        this.apiSource = apiSource;
         apiListing = null;
         resourcePath = "";
-        models = new HashMap<String, Model>();
         produces = new ArrayList<String>();
         consumes = new ArrayList<String>();
-        this.overriderConverter = overrideConverter;
+        this.logger = logger;
         this.parameterGenerator = new ParameterGenerator(logger);
+        this.parameterFilter = parameterFilter;
+        this.modelCollector = new ModelCollector(overrideConverter);
     }
 
     /**
@@ -157,7 +149,7 @@ public class SpringMvcApiReader {
                 scala.collection.immutable.List.fromIterator(JavaConversions.asScalaIterator(protocols.iterator())),
                 scala.collection.immutable.List.fromIterator(JavaConversions.asScalaIterator(authorizations.iterator())),
                 scala.collection.immutable.List.fromIterator(JavaConversions.asScalaIterator(apiDescriptions.iterator())),
-                generateModels(models), Option.apply(description), position);
+                generateModels(modelCollector.getModels()), Option.apply(description), position);
         return apiListing;
     }
 
@@ -232,7 +224,7 @@ public class SpringMvcApiReader {
             } else {
                 responseBodyName = (clazz.getSimpleName());
             }
-            addToModels(clazz);
+            modelCollector.addModel(clazz);
         }
         if (requestMapping.method() != null && requestMapping.method().length != 0) {
             method = requestMapping.method()[0].toString();
@@ -289,11 +281,20 @@ public class SpringMvcApiReader {
         List<Parameter> params = new ArrayList<Parameter>();
         for (int i = 0; i < annotations.length; i++) { //loops through parameters
             Class<?> parameterClass = m.getParameterTypes()[i];
-            addToModels(parameterClass);
-            Parameter parameter = parameterGenerator.generateParameter(new ParameterMetadata(parameterClass, annotations[i]));
-            params.add(parameter);
+            ParameterMetadata parameterMetadata = new ParameterMetadata(parameterClass, annotations[i]);
+            if (isNotIgnoredParameter(parameterMetadata)) {
+                modelCollector.addModel(parameterClass);
+                Parameter parameter = parameterGenerator.generateParameter(parameterMetadata);
+                params.add(parameter);
+            } else {
+                logger.info("Ignoring parameter " + parameterMetadata);
+            }
         }
         return params;
+    }
+
+    private boolean isNotIgnoredParameter(ParameterMetadata parameterMetadata) {
+        return !parameterFilter.isIgnoredParameter(parameterMetadata);
     }
 
     /**
@@ -310,7 +311,7 @@ public class SpringMvcApiReader {
                 if (apiResponse.response() == null || apiResponse.response().equals(java.lang.Void.class)) {
                     responseMessages.add(new ResponseMessage(apiResponse.code(), apiResponse.message(), Option.<String>empty()));
                 } else {
-                    addToModels(apiResponse.response());
+                    modelCollector.addModel(apiResponse.response());
                     responseMessages.add(new ResponseMessage(apiResponse.code(), apiResponse.message(), Option.apply(apiResponse.response().getSimpleName())));
                 }
             }
@@ -322,302 +323,7 @@ public class SpringMvcApiReader {
         return responseMessages;
     }
 
-    /**
-     * Generates a Model object for a Java class. Takes properties from either fields or methods.
-     * Recursion occurs if a ModelProperty needs to be modeled
-     *
-     * @param Class<?> clazz
-     * @return Model
-     */
-    private Model generateModel(Class<?> clazz) {
-        ApiModel apiModel;
-        ModelRef modelRef = null;
-        String modelDescription = "";
-        LinkedHashMap<String, ModelProperty> modelProps = new LinkedHashMap<String, ModelProperty>();
-        List<String> subTypes = new ArrayList<String>();
-
-        if (clazz.isAnnotationPresent(ApiModel.class)) {
-            apiModel = clazz.getAnnotation(ApiModel.class);
-            if (apiModel.description() != null) {
-                modelDescription = apiModel.description();
-            }
-        }
-
-        //<--Model properties from fields-->
-        int x = 0;
-
-        for (Field field : sortFields(clazz)) {
-            //Only use fields if they are annotated - otherwise use methods
-            XmlElement xmlElement;
-            ApiModelProperty amp;
-            Class<?> c;
-            String name;
-            boolean required = false;
-            String description = "";
-            if (!(field.getType().equals(clazz))) {
-                //if the types are the same, model will already be generated
-                modelRef = generateModelRef(clazz, field); //recursive IFF there is a generic sub-type to be modeled
-            }
-            if (field.isAnnotationPresent(XmlElement.class) ||
-                    field.isAnnotationPresent(ApiModelProperty.class)) {
-                if (field.getAnnotation(XmlElement.class) != null) {
-                    xmlElement = field.getAnnotation(XmlElement.class);
-                    required = xmlElement.required();
-                }
-                if (field.getAnnotation(ApiModelProperty.class) != null) {
-                    amp = field.getAnnotation(ApiModelProperty.class);
-                    if (!required) { //if required has already been changed to true, it was noted in XmlElement
-                        required = amp.required();
-                    }
-                    description = amp.value();
-                }
-
-                if (!(field.getType().equals(clazz))) {
-                    c = field.getType();
-                    name = field.getName();
-                    addToModels(c);
-                    subTypes.add(c.getSimpleName());
-                    modelProps.put(name, generateModelProperty(c, x, required, modelRef, description));
-                    x++;
-                }
-            }
-        }
-
-        //<--Model properties from methods-->
-        int i = 0;
-        for (Method m : sortMethodsGettersFirst(clazz)) {
-            boolean required = false;
-            String description = "";
-            ApiModelProperty amp;
-            //look for required field in XmlElement annotation
-
-            if (!(m.getReturnType().equals(clazz))) {
-                modelRef = generateModelRef(clazz, m); //recursive IFF there is a generic sub-type to be modeled
-            }
-
-            if (m.isAnnotationPresent(XmlElement.class)) {
-                XmlElement xmlElement = m.getAnnotation(XmlElement.class);
-                required = xmlElement.required();
-            } else if (m.isAnnotationPresent(JsonIgnore.class)
-                    || m.isAnnotationPresent(XmlTransient.class)) {
-                continue; //ignored fields
-            }
-
-            if (m.getAnnotation(ApiModelProperty.class) != null) {
-                amp = m.getAnnotation(ApiModelProperty.class);
-                required = amp.required();
-                description = amp.value();
-            }
-
-            //get model properties from methods
-            if (isGetterMethod(m)) {
-                String name = getPropertyNameFromMethodName(m);
-                if (!(m.getReturnType().equals(clazz))) {
-                    addToModels(m.getReturnType()); //recursive
-                }
-                if (!modelProps.contains(name)) {
-                    modelProps.put(name, generateModelProperty(m.getReturnType(), i, required, modelRef, description));
-                }
-            }
-            i++;
-        }
-
-        return new Model(clazz.getSimpleName(), clazz.getSimpleName(), clazz.getCanonicalName(), modelProps,
-                Option.apply(modelDescription), DEFAULT_OPTION, DEFAULT_OPTION,
-                scala.collection.immutable.List.fromIterator(JavaConversions.asScalaIterator(subTypes.iterator())));
-    }
-
-    private static Field[] sortFields(Class<?> clazz) {
-        Field[] sortedFields = clazz.getDeclaredFields();
-        Arrays.sort(sortedFields, new Comparator<Field>() {
-            @Override
-            public int compare(Field o1, Field o2) {
-                return o1.getName().compareTo(o2.getName());
-            }
-        });
-        return sortedFields;
-    }
-
-    private static Method[] sortMethodsGettersFirst(Class<?> clazz) {
-        Method[] ms = clazz.getMethods();
-        Arrays.sort(ms, new Comparator<Method>() {
-            @Override
-            public int compare(Method m1, Method m2) {
-                boolean m1IsGetter = isGetterMethod(m1);
-                boolean m2IsGetter = isGetterMethod(m2);
-                if (m1IsGetter) {
-                    if (m2IsGetter) {
-                        return getPropertyNameFromMethodName(m1).compareTo(getPropertyNameFromMethodName(m2));
-                    } else {
-                        return 1;
-                    }
-                } else {
-                    if (!m2IsGetter) {
-                        return m1.getName().compareTo(m2.getName());
-                    } else {
-                        return -1;
-                    }
-                }
-            }
-        });
-        return ms;
-    }
-
-    private static String getPropertyNameFromMethodName(Method method) {
-        String stripPrefix;
-        String methodName = method.getName();
-        if (methodName.startsWith("get") && methodName.length() > 3) {
-            stripPrefix = methodName.substring(3);
-        } else if (methodName.startsWith("is") && methodName.length() > 2) {
-            stripPrefix = methodName.substring(2);
-        } else {
-            stripPrefix = methodName;
-        }
-        return lowerCamelCase(stripPrefix);
-    }
-
-    private static String lowerCamelCase(String name) {
-        if (name.isEmpty()) {
-            return name;
-        }
-        String lowerFirstLetter = name.substring(0, 1).toLowerCase(); //convert to camel case
-        if (name.length() > 1) {
-            return lowerFirstLetter + name.substring(1);
-        } else {
-            return lowerFirstLetter;
-        }
-    }
-
-    private static boolean isGetterMethod(Method m) {
-        String name = m.getName();
-        return (name.startsWith("get") || name.startsWith("is")) && !(name.equals("getClass"));
-    }
-
-    /**
-     * Generates a ModelProperty for given model class. Supports String enumerations only.
-     *
-     * @param Class<?> clazz
-     * @param int      position
-     * @param boolean  required
-     * @param ModelRef modelRef
-     * @param String   description
-     * @return ModelProperty
-     */
-    private static ModelProperty generateModelProperty(Class<?> clazz, int position, boolean required, ModelRef modelRef,
-                                                String description) {
-        AllowableListValues allowed = null;
-        String name = clazz.getSimpleName();
-
-        if (!(isModel(clazz))) {
-            name = name.toLowerCase();
-        }
-        //check for enumerated values - currently strings only
-        //TODO: support ranges
-        if (clazz.isEnum()) {
-            List<String> enums = new ArrayList<String>();
-            for (Object obj : clazz.getEnumConstants()) {
-                enums.add(obj.toString());
-            }
-            allowed = new AllowableListValues(
-                    scala.collection.immutable.List.fromIterator(JavaConversions.asScalaIterator(enums.iterator())), "LIST");
-        }
-
-        return new ModelProperty(name,
-                clazz.getCanonicalName(), position, required, Option.apply(description), allowed,
-                Option.apply(modelRef));
-    }
-
-    /**
-     * Generates a model reference based on a method
-     *
-     * @param Class<?> clazz
-     * @param Method   m
-     * @return ModelRef
-     */
-    private ModelRef generateModelRef(Class<?> clazz, Method m) {
-        ModelRef modelRef = null; //can be null
-        if (Collection.class.isAssignableFrom(m.getReturnType()) || m.getReturnType().equals(ResponseEntity.class)
-                || m.getReturnType().equals(JAXBElement.class)) {
-            Class<?> c = getGenericSubtype(m.getReturnType(), m.getGenericReturnType());
-            if (isModel(c) && !(c.equals(clazz))) {
-                addToModels(c);
-                modelRef = new ModelRef(c.getSimpleName(), Option.apply(c.getSimpleName()),
-                        Option.apply(c.getSimpleName()));
-            } else {
-                modelRef = new ModelRef(c.getSimpleName().toLowerCase(), DEFAULT_OPTION, DEFAULT_OPTION);
-            }
-        }
-        return modelRef;
-    }
-
-    /**
-     * Generates a model reference based on a field
-     *
-     * @param Class<?> clazz
-     * @param Field    f
-     * @return ModelRef
-     */
-    private ModelRef generateModelRef(Class<?> clazz, Field f) {
-        ModelRef modelRef = null;
-        if (Collection.class.isAssignableFrom(f.getType()) || f.getType().equals(ResponseEntity.class)
-                || f.getType().equals(JAXBElement.class)) {
-            Class<?> c = getGenericSubtype(f.getType(), f.getGenericType());
-            if (isModel(c) && !(c.equals(clazz))) {
-                addToModels(c);
-                modelRef = new ModelRef(c.getSimpleName(), Option.apply(c.getSimpleName()),
-                        Option.apply(c.getSimpleName()));
-            } else {
-                modelRef = new ModelRef(c.getSimpleName().toLowerCase(), DEFAULT_OPTION, DEFAULT_OPTION);
-            }
-        }
-        return modelRef;
-    }
-
-    //-------------Helper Methods------------//
-
-    private static Class<?> getGenericSubtype(Class<?> clazz, Type t) {
-        if (!(clazz.getName().equals("void") || t.toString().equals("void"))) {
-            try {
-                ParameterizedType paramType = (ParameterizedType) t;
-                Type[] argTypes = paramType.getActualTypeArguments();
-                if (argTypes.length > 0) {
-                    return (Class<?>) argTypes[0];
-                }
-            } catch (ClassCastException e) {
-                //FIXME: find out why this happens to only certain types
-            }
-        }
-        return clazz;
-    }
-
-    private void addToModels(Class<?> clazz) {
-        if (isModel(clazz) && !(models.containsKey(clazz.getSimpleName()))) {
-            //put the key first in models to avoid stackoverflow
-            models.put(clazz.getSimpleName(), new Model(null, null, null, null, null, null, null, null));
-            scala.collection.mutable.HashMap<String, Option<Model>> overriderMap = this.overriderConverter.overrides();
-            if (overriderMap.contains(clazz.getCanonicalName())) {
-                Option<Option<Model>> m = overriderMap.get(clazz.getCanonicalName());
-                models.put(clazz.getSimpleName(), m.get().get());
-            } else {
-                models.put(clazz.getSimpleName(), generateModel(clazz));
-            }
-        }
-    }
-
-    private static boolean isModel(Class<?> clazz) {
-        try {
-            for (String str : RESERVED_PACKAGES) {
-                if (clazz.getPackage().getName().contains(str)) {
-                    return false;
-                }
-            }
-            return !SwaggerSpec.baseTypes().contains(clazz.getSimpleName().toLowerCase());
-        } catch (NullPointerException e) { //null pointer for package names - wouldn't model something without a package. skip
-            return false;
-        }
-    }
-
-    private Option<scala.collection.immutable.Map<String, Model>> generateModels(HashMap<String, Model> javaModelMap) {
+    private Option<scala.collection.immutable.Map<String, Model>> generateModels(Map<String, Model> javaModelMap) {
         return Option.apply(Utils.toScalaImmutableMap(javaModelMap));
     }
